Trait tidechain_service::ClientHandle

pub trait ClientHandle {
    fn execute_with<T>(&self, t: T) -> <T as ExecuteWithClient>::Output
    where
        T: ExecuteWithClient;
}

A handle to a Tidechain client instance.

The Tidechain service supports multiple different runtimes (Lagoon, Tidechain itself, etc). As each runtime has a specialized client, we need to hide them behind a trait. This is this trait.

When wanting to work with the inner client, you need to use execute_with.

See ExecuteWithClient for more information.

Required Methods

fn execute_with<T>(&self, t: T) -> <T as ExecuteWithClient>::Outputwhere
T: ExecuteWithClient,

Execute the given something with the client.