Trait tidechain_service::IdentifyVariant

pub trait IdentifyVariant {
    fn is_tidechain(&self) -> bool;
    fn is_lagoon(&self) -> bool;
    fn is_dev(&self) -> bool;
}

Can be called for a Configuration to check if it is a configuration for the Lagoon network.

Required Methods

fn is_tidechain(&self) -> bool

Returns if this is a configuration for the Tidechain network.

fn is_lagoon(&self) -> bool

Returns if this is a configuration for the Lagoon network.

fn is_dev(&self) -> bool

Returns true if this configuration is for a development network.