Crate tidechain_service

Re-exports

pub use chain_spec::LagoonChainSpec;

pub use chain_spec::TidechainChainSpec;

pub use tidechain_runtime;

pub use lagoon_runtime;

Modules

chain_spec

Macros

tidefi_staking_genesis

tidefi_sunrise_pool_genesis

Structs

LagoonExecutorDispatch

NewChainOps

NewFull

TidechainExecutorDispatch

Enums

Client

A client instance of Tidechain.

Error

Traits

AbstractClient

Trait that abstracts over all available client implementations.

ChainSpec

Common interface of a chain specification.

ClientHandle

A handle to a Tidechain client instance.

ExecuteWithClient

Execute something with the client instance.

IdentifyVariant

Can be called for a Configuration to check if it is a configuration for the Lagoon network.

RuntimeApiCollection

A set of APIs that Tidechain-like runtimes must implement.

Functions

build_full

new_chain_ops

Builds a new object suitable for chain operations.

new_full

Type Definitions

FullBackend

FullClient

FullGrandpaBlockImport

FullSelectChain