Crate tidechain_service

Re-exports

pub use tidechain_runtime;

pub use hertel_runtime;

Modules

chain_spec

Structs

HertelExecutorDispatch

NewChainOps

NewFull

TidechainExecutorDispatch

Enums

Client

A client instance of Tidechain.

Error

Traits

AbstractClient

Trait that abstracts over all available client implementations.

ClientHandle

A handle to a Polkadot client instance.

ExecuteWithClient

Execute something with the client instance.

IdentifyVariant

Can be called for a Configuration to check if it is a configuration for the Hertel network.

RuntimeApiCollection

A set of APIs that Tidechain-like runtimes must implement.

Functions

build_full

new_chain_ops

Builds a new object suitable for chain operations.

new_full

Type Definitions

FullBackend

FullClient

FullGrandpaBlockImport

FullSelectChain